+++ lib/polish_geeks/dev_tools/validators/rubocop.rb
@@ -1,0 +1,12 @@
+module PolishGeeks
+  module DevTools
+    module Validators
+      # It verifies if Rubocop is enabled
+      class Rubocop < Base
+        def valid?
+          PolishGeeks::DevTools.config.rubocop == true
+        end
+      end
+    end
+  end
+end